version_03/requirements.txt
# Core dependencies
playwright==1.41.1
python-dotenv==1.0.0

# Web automation and HTTP
requests==2.31.0
aiohttp==3.9.1
urllib3==2.1.0

# HTML parsing
beautifulsoup4==4.12.2
html5lib==1.1

# Async support
asyncio-throttle==1.0.2